feat(metrics): count every playback start as view with fractional loops
View = playback start per 2026-08-13 spec, not >=1s. Every session
that starts tracking now emits one view even if N=0 loops (scroll-past,
still-buffering, quick swipe all count). Fractional loops preserved so
the median 0.75 case does not round to zero. Retry service and publisher
no longer gate on duration; widget computes fractional loops from
watchDuration/totalDuration. Settles open decision as fractional loops.

Fixes undercount that left 73% of interactors absent from view counts
and suppressed creator celebration. Display path remains raw sums with
publicLoopCountFloor=1000; ranking remains damped via log1p(loops)*0.5.

+++ mobile/test/widgets/video_feed_item/feed_videos_test.dart
@@ -131,7 +131,7 @@ class _NoopAnalyticsService extends AnalyticsService {
     required String eventType,
     Duration? watchDuration,
     Duration? totalDuration,
-    int? loopCount,
+    double? loopCount,
     bool? completedVideo,
     ViewTrafficSource trafficSource = ViewTrafficSource.unknown,
     String? sourceDetail,
